Make use of 'source_hash_update'
From salt v2016.3.0 we can use 'source_hash_update' to automatically
update the contents of the extracted archive directory when the hash of
the source changes. This means we can use salt to upgrade or change the
version of gogs installed just by updating the pillar 'binary_archive'
and 'binary_archive_hash' values.