Manage inbuilt log file rotation using pillar data
Gogs provides an option to use its inbuilt log rotation to rotate the
'gogs.log' file. The gogs formula uses logrotate to do log rotation so
the inbuilt functionality is disabled by default.

Currently this option only affects the rotation of the 'gogs.log' file,
but this may change in the future. See:

    https://github.com/gogits/gogs/issues/3413