Move config file values under 'config' in pillar
The 'lookup' section in the pillar data should really only be used to
modify how Gogs is installed, not how it is configured. Settings for
controlling how Gogs is configured should now be under the 'config'
section in the pillar data.

The upshot of this is that most of the default config values are managed
in the app.ini.jinja file which should generally make it easier to
manage that file.