Don't use 'source_hash_update' option with 'gogs-binary-package'
The 'source_hash_update' option does not work with remote sources.
Instead of having some kludgey hack to try and determine if the source
is a remote on or not and use it conditionally, it is much simpler to
remove it.

It seems like this feature may be fixed in the develop branch of Salt,
so we may look at adding it back in once there is a release of Salt that
supports using 'source_hash_update' with remote sources.