Use `enforce_toplevel` for Salt >=2016.11.0
We are using `--strip-components=1` to remove the top-level directory
from the extracted archive, but from v2016.11.2 salt complains loudly
about a potential 'tar-bomb'.11.2. The `enforce_toplevel` option was
introduced in v2016.11.0 and can be used to avoid this error, so we may
as well use it all versions of salt that support the option.